Simplify Controller by moving code into PersonService method 'searchByNames'
